Move more code over into learning script
